Crate nvimpam_lib

The companion library to the nvimpam binary.

Modules

card
A datastructure to represent a Card of the FEM-Solver Pamcrash.

carddata
This module contains all the global static Card instances

event
The events that nvimpam needs to accept and deal with. They're sent by the NeovimHandler to the main loop.

folds
This module provides the FoldList struct to manage folds in a buffer. It carries both level 1 folds as well as level 2 folds (i.e. folds that contain folds of level 1). All functions that insert/remove/modify folds operate on level 1 folds, the only thing to be done for the level 2 folds is regenerating them in full from the level 1 folds.

handler
The handler for the rpc events sent by neovim_lib. Note that this is excuted in another thread, so we use a Sender<Event> to send the parsed event data to the main thread.

lines
This module holds the datastructure for the Lines of the buffer.

nocommentiter
This module holds NoCommentIter, the central datastructure for the folding functionality of nvimpam.

skipresult
This module holds datastructures for the various skip_* methods of the NoCommentIter