Module nvimpam_lib::card

A datastructure to represent a Card of the FEM-Solver Pamcrash.

There is a 1-1 correspondence between a Card and a Keyword, but they are used differently. A Keyword is used as a parsing result for strings all over the place, while Card is used to define global static values the represent the concrete CARD entities of Pamcrash.

The definitions of the global static values can be found in the carddata module.

Modules

cell
Elements of an input line

ges
The General Entity Selection scheme of Pamcrash.

keyword
This module provides the Keyword enum to classify lines according to what card type they belong to. The terms "Keyword" and "Card" are lingo from the FEM solver Pamcrash, but generally used among FEM solvers.

line
An enum to classify the several types of lines that can occur inside a card of a Pamcrash input file. Might not really be a line (see GES, rather than zero or more lines.

Structs

Card
A card consists of severals Lines, and starts with a given Keyword. If ownfold is true, than each card of this type will get an own fold. Otherwise, all adjacent cards of that type are gathered into one fold.