Enum nvimpam_lib::card::keyword::Keyword

pub enum Keyword {
    Comment,
    Node,
    Cnode,
    Mass,
    Nsmas,
    Nsmas2,
    Solid,
    Hexa20,
    Pent15,
    Penta6,
    Tetr10,
    Tetr4,
    Bshel,
    Tshel,
    Shell,
    Shel6,
    Shel8,
    Membr,
    Beam,
    Sprgbm,
    Bar,
    Spring,
    Joint,
    Kjoin,
    Mtojnt,
    Sphel,
    Sphelo,
    Gap,
    Impma,
    Elink,
    Llink,
    Slink,
    Plink,
    Tied,
    PartSolid,
    PartBshel,
    PartTetra,
    PartSphel,
    PartCos3d,
    PartTshel,
    PartShell,
    PartMembr,
    PartBar,
    PartBeam,
    PartSpring,
    PartSprgbm,
    PartMbspr,
    PartJoint,
    PartKjoin,
    PartMbkjn,
    PartMtojnt,
    PartTied,
    PartSlink,
    PartElink,
    PartLlink,
    PartPlink,
    PartGap,
    Mtoco,
    Otmco,
    Rbody0,
    Rbody1,
    Rbody2,
    Rbody3,
    Group,
}

An enum to denote the several types of cards a line might belong to.

Variants
CommentNodeCnodeMassNsmasNsmas2SolidHexa20Pent15Penta6Tetr10Tetr4BshelTshelShellShel6Shel8MembrBeamSprgbmBarSpringJointKjoinMtojntSphelSpheloGapImpmaElinkLlinkSlinkPlinkTiedPartSolidPartBshelPartTetraPartSphelPartCos3dPartTshelPartShellPartMembrPartBarPartBeamPartSpringPartSprgbmPartMbsprPartJointPartKjoinPartMbkjnPartMtojntPartTiedPartSlinkPartElinkPartLlinkPartPlinkPartGapMtocoOtmcoRbody0Rbody1Rbody2Rbody3Group
Methods

impl Keyword

pub fn parse(s: &[u8]) -> Option<Keyword>

Parse a string to determine if it starts with the keyword of a card.