Struct nvimpam_lib::folds::FoldList

pub struct FoldList { /* fields omitted */ }

Holds the fold data of the buffer. A fold has the following data: Linenumbers start, end (indexed from 0), and a Keyword.

Methods

impl FoldList

pub fn new() -> FoldList

Create a new FoldList. There does not seem to be a way to create one with a predetermined capacity.

pub fn clear(&mut self)

Clear FoldList, by clearing the BTreeMap's individually

pub fn checked_insert(
&mut self,
start: u64,
end: u64,
kw: Keyword
) -> Result<(), Error>

Insert a level 1 fold ([start, end], Keyword) into the FoldList. If end < start, we silently return. Otherwise, we call the internal insert function that returns an error if the fold is already in the list. In that case, it needs to be removed beforehand.

pub fn remove(&mut self, start: u64, end: u64) -> Result<(), Error>

Remove a level 1 fold [start, end] from the foldlist. Only checks if the fold is in the FoldList, and returns an error otherwise.

pub fn recreate_all(
&mut self,
keywords: &[Option<Keyword>],
lines: &[Line]
) -> Result<(), Error>

Remove all the entries from the FoldList, and iterate over lines to populate it with new ones. Then recreate the level 2 folds.

pub fn resend_all(&self, nvim: &mut Neovim) -> Result<(), Error>

Delete all folds in nvim, and create the ones from the FoldList.

pub fn to_vec(&self, level: u8) -> Vec<(u64, u64, Keyword)>

Copy the elements of a FoldList of the given level into a Vec, containing the tuples (start, end, Keyword)

pub fn add_folds(
&mut self,
keywords: &[Option<Keyword>],
lines: &[Line]
) -> Result<(), Error>

Parse an array of Option<Keyword>s into a FoldList. The foldlist is cleared as a first step.

Creates only level 1 folds. Depending on the ownfold parameter in the definition of the card in the carddata module, each card will be in an own fold, or several adjacent (modulo comments) cards will be subsumed into a fold.