Module nvimpam_lib::handler

The handler for the rpc events sent by neovim_lib. Note that this is excuted in another thread, so we use a Sender<Event> to send the parsed event data to the main thread.

Structs

NeovimHandler
The handler containing the sending end of a channel. The receiving end is the main event loop.

Functions

last_arg
Helper function to get the last argument of a Vec<Value> or return an error message

parse_bool
Parse a neovim_lib::Value into a bool

parse_buf
Parse a neovim_lib::Value into a neovim_lib::Buffer. This cannot fail, but if the Value was not obtained from the rpc api, this will probably not be a valid buffer to send commands to.

parse_i64
Parse a neovim_lib::Value into a i64

parse_u64
Parse a neovim_lib::Value into a u64

parse_vecstr
Parse a neovim_lib::Value into a Vec<String>. Note that this method takes ownership of the value so it does not need to copy out the contained strings